Added files to test slew rate
This is not complete, but this is an updated version of how to calculate the slew rate. The main variables are included in this version, but still need to be manipulated in an easy way to call them within a short, but reasonable time frame for the driver to see.